(fix): refactor types
