+++ design_system/components/molecules/ExternalLink/ExternalLink.tsx
@@ -13,9 +13,7 @@ export function ExternalLink({ href, ...rest }: Props) {
       href={href as any}
       onPress={async (event) => {
         if (Platform.OS !== 'web') {
-          // Prevent the default behavior of linking to the default browser on native.
           event.preventDefault();
-          // Open the link in an in-app browser.
           await openBrowserAsync(href);
         }
       }}
