+++ design_system/components/molecules/TabIcon/TabIcon.tsx
@@ -3,7 +3,7 @@ import { IconSymbol } from '@/design_system/ui/vendors';
 import { IconSymbolName } from '@/design_system/ui/vendors/IconSymbol';
 
 interface TabIconProps {
-  name: IconSymbolName; // Asegurar que solo acepte nombres válidos
+  name: IconSymbolName; 
   color: string;
   isLoading: boolean;
 }
@@ -11,7 +11,7 @@ interface TabIconProps {
 export const TabIcon: React.FC<TabIconProps> = ({ name, color, isLoading }) => (
   <IconSymbol 
     size={28} 
-    name={name} // Ya es del tipo correcto
+    name={name} 
     color={color}
     style={{ opacity: isLoading ? 0.5 : 1 }}
   />
