+++ hooks/db/useUserDb.ts
@@ -1,21 +1,14 @@
 import { useState, useEffect, useCallback } from 'react';
-import { db } from '../../database/db';
-import { users } from '../../database/schema';
+import { db } from '@/database/db';
+import { users } from '@/database/schema';
 import { eq } from 'drizzle-orm';
-
-export interface User {
-  id: string;
-  name: string;
-  avatar: string;
-  status: 'online' | 'offline' | 'away';
-}
+import { User } from '@/types/User';
 
 export function useUserDb() {
   const [allUsers, setAllUsers] = useState<User[]>([]);
   const [currentUser, setCurrentUser] = useState<User | null>(null);
   const [loading, setLoading] = useState(true);
 
-  // Validar que `db` esté disponible antes de usarlo
   useEffect(() => {
     if (!db) {
       console.error("Database is not initialized.");
@@ -26,7 +19,7 @@ export function useUserDb() {
     const loadUsers = async () => {
       try {
         const usersData = await db.select().from(users);
-        console.log('Users data:', usersData);
+
 
         if (!usersData || usersData.length === 0) {
           throw new Error('No data returned from the database');
@@ -51,14 +44,13 @@ export function useUserDb() {
       }
 
       const user = await db.select().from(users).where(eq(users.id, userId));
-      console.log('User data:', user); // Log the data returned from the database
+
 
       if (!user || user.length === 0) {
         console.error('No user found with the given ID');
         return false;
       }
 
-      // Ensure user[0] has the expected properties
       const { id, name, avatar, status } = user[0];
       if (!id || !name || !avatar || !status) {
         console.error('User data is missing required properties');
