+++ hooks/useUser.ts
@@ -1,6 +1,4 @@
-import { useUserDb, User } from './db/useUserDb';
-
-export { User };
+import { useUserDb } from './db/useUserDb';
 
 export function useUser() {
   const {
