+++ types/User.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: string;
+  name: string;
+  avatar: string;
+  status: 'online' | 'offline' | 'away';
+}
